// apps/live-preview-react-demo/provider/universalText.d.ts
import React from "react";
import { IGenUniversalText } from "../utils/types_gen";
export declare const useUniversalText: () => Omit<IGenUniversalText, "__typename">;
export declare const UniversalTextProvider: React.FC<React.PropsWithChildren<{
    universalText: IGenUniversalText;
}>>;
//# sourceMappingURL=universalText.d.ts.map