// apps/live-preview-react-demo/utils/types_gen.d.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    DateTime: any;
    JSON: any;
};
export type IGenAsset = {
    __typename: "Asset";
    _meta?: Maybe<IGenCaisyDocument_Meta>;
    author?: Maybe<Scalars["String"]>;
    copyright?: Maybe<Scalars["String"]>;
    description?: Maybe<Scalars["String"]>;
    dominantColor?: Maybe<Scalars["String"]>;
    id?: Maybe<Scalars["ID"]>;
    keywords?: Maybe<Scalars["String"]>;
    originType?: Maybe<Scalars["String"]>;
    originalName?: Maybe<Scalars["String"]>;
    src?: Maybe<Scalars["String"]>;
    title?: Maybe<Scalars["String"]>;
};
// ...
export type IGenCaisyDocument_Meta = {
    __typename: "CaisyDocument_Meta";
    createdAt?: Maybe<Scalars["DateTime"]>;
    id?: Maybe<Scalars["ID"]>;
    locale?: Maybe<Scalars["String"]>;
    locales?: Maybe<Array<Maybe<Scalars["String"]>>>;
    publishedAt?: Maybe<Scalars["DateTime"]>;
    updatedAt?: Maybe<Scalars["DateTime"]>;
};